+++ packages/npm/lib/react/checkbox.js
@@ -35,6 +35,8 @@ var Checkbox = function (_React$Component) {
   babelHelpers.createClass(Checkbox, [{
     key: 'render',
     value: function render() {
+      var _this2 = this;
+
       var _props = this.props,
           children = _props.children,
           className = _props.className,
@@ -61,7 +63,9 @@ var Checkbox = function (_React$Component) {
           'label',
           null,
           _react2.default.createElement('input', {
-            ref: 'inputEl',
+            ref: function ref(el) {
+              _this2.inputElRef = el;
+            },
             type: 'checkbox',
             autoFocus: autoFocus,
             checked: checked,